+++ R/sendrecv.R
@@ -252,50 +252,53 @@ recv_aio <- function(socket,
     return(invisible(`class<-`(aio, "errorValue")))
   }
   env <- `class<-`(new.env(), "recvAio")
-  data <- raw <- resolv <- NULL
+  data <- raw <- NULL
+  unresolv <- TRUE
   if (keep.raw) {
     makeActiveBinding(sym = "raw", fun = function(x) {
-      if (is.null(resolv)) {
+      if (unresolv) {
         res <- .Call(rnng_aio_get_msg, aio)
         missing(res) && return(.Call(rnng_aio_unresolv))
         is.integer(res) && {
-          data <<- raw <<- resolv <<- `class<-`(res, "errorValue")
+          data <<- raw <<- `class<-`(res, "errorValue")
+          unresolv <<- FALSE
           logerror(res)
-          return(invisible(resolv))
+          return(invisible(data))
         }
         on.exit(expr = {
           raw <<- res
-          resolv <<- 0L
+          unresolv <<- FALSE
           return(res)
         })
         data <- decode(con = res, mode = mode)
         on.exit()
         raw <<- res
         data <<- data
-        resolv <<- 0L
+        unresolv <<- FALSE
       }
       raw
     }, env = env)
   }
   makeActiveBinding(sym = "data", fun = function(x) {
-    if (is.null(resolv)) {
+    if (unresolv) {
       res <- .Call(rnng_aio_get_msg, aio)
       missing(res) && return(.Call(rnng_aio_unresolv))
       is.integer(res) && {
-        data <<- raw <<- resolv <<- `class<-`(res, "errorValue")
+        data <<- raw <<- `class<-`(res, "errorValue")
+        unresolv <<- FALSE
         logerror(res)
-        return(invisible(resolv))
+        return(invisible(data))
       }
       on.exit(expr = {
         data <<- res
-        resolv <<- 0L
+        unresolv <<- FALSE
         return(res)
       })
       data <- decode(con = res, mode = mode)
       on.exit()
       if (keep.raw) raw <<- res
       data <<- data
-      resolv <<- 0L
+      unresolv <<- FALSE
     }
     data
   }, env = env)